requirements.txt
BeautifulSoup==3.2.0
Django==1.3.1
Markdown==2.0.3
PIL==1.1.7
South==0.7.3
Whoosh==2.3.0
argparse==1.2.1
django-appmedia==1.0.1
django-blog-zinnia==0.9
django-classy-tags==0.3.4.1
django-cms==2.2
django-cms-taggit==0.1
django-mptt==0.5.0
django-sekizai==0.5
django-tagging==0.3.1
django-taggit==0.9.3
django-xmlrpc==0.1.3
haystack==0.13
html5lib==0.90
pyparsing==1.5.6
python-ptrace==0.6.3
wsgiref==0.1.2
